Move {un}flatten to utils
